Add files via upload
A fun project that I came up with one night, and still working on it so I can make it run perfectly. 
I see it as a very simplistic version of a grocery delivery/grocery store/recipe app, where we can help the user to come up with a simple recipe to cook, from the ingredients that they have at home.
The app will ask two inputs from the user and then generates two random values into the final print sentence. The random commit keeps the app more interesting, surprising and engaging.