Update recipe_helper.py
A fun project I came up with one night, trying to help the user to come up with suggestions of an easy recipe to make out of ingredients the user can find from home. I can see this as a simplistic version of an app/attribute of an app for a grocery store/food delivery/recipe application to intuitively help customer with creating recipes from mainly from the ingredients found at home and help save money and reduce food waste.