Add files via upload
One of the tasks at HyperionDev DataScience bootcamp to practice the use of lists and dictionaries. In this task we were asked to first create a list of all the products sold in the cafe, and then two dictionaries, one for the total stock for each item and another one for a price for each item. At the end of the program, we calculated the sum of the total stock.