Add files via upload
This task at HyperionDev Data Science Bootcamp was about practicing Logical Programming - Operators. In this program, we asked the user to enter the time for each part of the triathlon and then calculate the total time.

At the end, depending on the total time, the program told which award each contestant would receive.