Add files via upload
This is a task regarding Natural Language Processing. In this HyperionDev task, we were to create a list of garden path sentences - a sentence characterized by a surprising twist at the end,  having the reader revisit the beginning and reassess the sentence's concept upon a second reading.

In the list, I tokenized each sentence and performed named entity recognition, examining how spaCy categorizes each sentence.

After that I printed a list of entities and their meanings.

With some of the sentences, the entity was completely missing.

I intentionally wanted to test whether spaCy recognized the difference between 'June' as a month and 'June' as a woman's name and frankly it did not. Also, it did not recognize Turkey as a country.